backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.94.0",
    "fastapi>=0.135.3",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
